Ajustes

+++ resources/views/users/edit.blade.php
@@ -57,7 +57,7 @@
                             </div>
                         {!! Form::close() !!}
 
-                        {!! Form::model($user, ['method' => 'PATCH','route' => ['users.reset', $user->id]]) !!}
+                        {!! Form::model($user, ['method' => 'PATCH','route' => ['users.reset', $user->id], 'class' => 'confirmation']) !!}
                             <div class="col-xs-12 col-sm-12 col-md-12">
                                 <div class="form-group">
                                     {!! Form::button('<i class="fa fa-reply"> Reestablecer Contraseña</i>', ['type' => 'submit', 'class' => 'btn btn-danger']) !!}
@@ -65,7 +65,7 @@
                             </div>
                         {!! Form::close() !!}
 
-                        {!! Form::model($user, ['method' => 'DELETE','route' => ['questions.destroy', $user->id]]) !!}
+                        {!! Form::model($user, ['method' => 'DELETE','route' => ['questions.destroy', $user->id], 'class' => 'confirmation']) !!}
                         <div class="col-xs-12 col-sm-12 col-md-12">
                             <div class="form-group">
                                 {!! Form::button('<i class="fa fa-trash"> Eliminar Preguntas de Seguridad</i>', ['type' => 'submit', 'class' => 'btn btn-danger']) !!}
@@ -80,4 +80,27 @@
             </div>
         </div>
     </section>
+@endsection
+
+@section('scripts')
+
+    <script>
+        $('.confirmation').submit(function(e){
+            e.preventDefault();
+
+            Swal.fire({
+            title: '¿Estás seguro?',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: '¡Sí, estoy seguro!'
+            }).then((result) => {
+            if (result.value) {
+                this.submit();
+            }
+            })
+        });
+    </script>
+
 @endsection
